requirements.txt
fastapi
uvicorn
python-dotenv
jinja2
python-multipart
PyPDF2
openai
google-generativeai